fix(assertions): keep grader projection internal
