+++ src/assertions/agentRubric.ts
@@ -1,7 +1,9 @@
 import { matchesAgentRubric } from '../matchers/agent';
 import { isGraderFailure } from '../matchers/llmGrading';
-import { type AssertionParams, type GradingResult, getGraderVars } from '../types/index';
 import invariant from '../util/invariant';
+import { getGraderVars } from './utils';
+
+import type { AssertionParams, GradingResult } from '../types/index';
 
 export const handleAgentRubric = async ({
   assertion,
