+++ src/assertions/contextFaithfulness.ts
@@ -1,7 +1,9 @@
 import { matchesContextFaithfulness } from '../matchers/rag';
-import { type AssertionParams, type GradingResult, getGraderVars } from '../types/index';
 import invariant from '../util/invariant';
 import { resolveContext } from './contextUtils';
+import { getGraderVars } from './utils';
+
+import type { AssertionParams, GradingResult } from '../types/index';
 
 /**
  * Handles context-faithfulness assertions by evaluating whether the LLM output
