+++ src/assertions/llmRubric.ts
@@ -1,6 +1,8 @@
 import { isGraderFailure, matchesLlmRubric } from '../matchers/llmGrading';
-import { type AssertionParams, type GradingResult, getGraderVars } from '../types/index';
 import invariant from '../util/invariant';
+import { getGraderVars } from './utils';
+
+import type { AssertionParams, GradingResult } from '../types/index';
 
 export const handleLlmRubric = async ({
   assertion,
