+++ src/external/assertions/deepeval.ts
@@ -3,16 +3,33 @@
 
 import { callProviderWithContext, getAndCheckProvider } from '../../matchers/providers';
 import { getDefaultProviders } from '../../providers/defaults';
-import { type AssertionParams, type GradingResult, getGraderVars } from '../../types/index';
 import invariant from '../../util/invariant';
 import { extractJsonObjects } from '../../util/json';
 import { accumulateTokenUsage, createEmptyTokenUsage } from '../../util/tokenUsageUtils';
 import { ConversationRelevancyTemplate } from '../matchers/conversationRelevancyTemplate';
 import { matchesConversationRelevance } from '../matchers/deepeval';
 
+import type { AssertionParams, GradingResult } from '../../types/index';
 import type { Message } from '../matchers/deepeval';
 
 const DEFAULT_WINDOW_SIZE = 5;
+
+function getConversationGraderVars(
+  assertion: AssertionParams['assertion'],
+  vars: AssertionParams['test']['vars'],
+) {
+  if (assertion.graderVars === undefined) {
+    return vars;
+  }
+  if (!vars) {
+    return {};
+  }
+  return Object.fromEntries(
+    assertion.graderVars
+      .filter((name) => name !== 'messages' && Object.prototype.hasOwnProperty.call(vars, name))
+      .map((name) => [name, vars[name]]),
+  );
+}
 
 export const handleConversationRelevance = async ({
   assertion,
@@ -51,7 +68,7 @@ export const handleConversationRelevance = async ({
     const result = await matchesConversationRelevance(
       windowMessages,
       1.0, // Use 1.0 threshold for individual windows
-      getGraderVars(assertion, test.vars),
+      getConversationGraderVars(assertion, test.vars),
       test.options,
       providerCallContext,
     );
